perf(masir): introduce class and hwnd pair caches
This commit introduces caches for class names and pairs of hwnds which
refer to the same application.

Once again, these caches will be used to reduce the number of syscalls
and file reads, and in the case of the hwnd pair cache, will allow masir
to exit an iteration early if it has already discovered two hwnds with
different classes that refer to the same application.